// client/src/store/user/consts.ts
export enum UserActionTypes {
    SIGNIN_REQUEST = 'SIGNIN_REQUEST',
    SIGNIN_SUCCESS = 'SIGNIN_SUCCESS',
    SIGNIN_FAILED = 'SIGNIN_FAILED',

    SIGNUP_REQUEST = 'SIGNUP_REQUEST',
    SIGNUP_SUCCESS = 'SIGNUP_SUCCESS',
    SIGNUP_FAILED = 'SIGNUP_FAILED',

    USER_DETAILS_REQUEST = 'USER_DETAILS_REQUEST',
    USER_DETAILS_SUCCESS = 'USER_DETAILS_SUCCESS',
    USER_DETAILS_FAILED = 'USER_DETAILS_FAILED',

    USER_UPDATE_REQUEST = 'USER_UPDATE_REQUEST',
    USER_UPDATE_SUCCESS = 'USER_UPDATE_SUCCESS',
    USER_UPDATE_FAILED = 'USER_UPDATE_FAILED',

    USER_LIST_REQUEST = 'USER_LIST_REQUEST',
    USER_LIST_SUCCESS = 'USER_LIST_SUCCESS',
    USER_LIST_FAILED = 'USER_LIST_FAILED',
    USER_LIST_CLEAR = 'USER_LIST_CLEAR',

    USER_DELETE_REQUEST = 'USER_DELETE_REQUEST',
    USER_DELETE_SUCCESS = 'USER_DELETE_SUCCESS',
    USER_DELETE_FAILED = 'USER_DELETE_FAILED',

    USER_SIGNOUT = 'USER_SIGNOUT',
    CLEAR_ERROR = 'CLEAR_ERROR',
}